+++ src/models/recommendation_history.py
@@ -6,6 +6,8 @@ from mongoengine import StringField, DateTimeField, FloatField, \
 from commons.enum import ListEnum
 from models.base_model import BaseModel
 
+RESOURCE_TYPE_INSTANCE = 'INSTANCE'
+RESOURCE_TYPE_GROUP = 'GROUP'
 
 class FeedbackStatusEnum(ListEnum):
     APPLIED = 'APPLIED'
@@ -52,8 +54,10 @@ class RecommendationTypeEnum(ListEnum):
                 cls.ACTION_SCALE_DOWN,
                 cls.ACTION_SPLIT]
 
+
 class RecommendationHistory(BaseModel):
-    instance_id = StringField(null=True)
+    resource_id = StringField(null=True)
+    resource_type = StringField(null=True)
     job_id = StringField(null=True)
     customer = StringField(null=True)
     tenant = StringField(null=True)
@@ -69,14 +73,13 @@ class RecommendationHistory(BaseModel):
     feedback_status = EnumField(FeedbackStatusEnum, null=True)
     last_metric_capture_date = DateTimeField(null=True)
 
-    dto_skip_attrs = ['savings', 'instance_meta']
     meta = {
         'indexes': [
-            'instance_id',
+            'resource_id',
             'customer',
-            ('instance_id', 'job_id'),
+            ('resource_id', 'job_id'),
             {
-                'fields': ['instance_id', 'added_at', 'recommendation_type'],
+                'fields': ['resource_id', 'added_at', 'recommendation_type'],
                 'unique': True
             },
             {