+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.6.0] - 2024-02-05
+* implement storing resource group recommendations in db
+* implement resource group cooldown: past resource group recommendation will be reused for the period of cooldown
+
 ## [3.5.7] - 2024-02-01
 * skip group resources without latest metrics available from AUTO_SCALING recommendation processing
 * improve group policy threshold validation